Split workspace preparation and building into independent stages
This will allow us to re-do a compilation from a pre-created workspace.
